+++ src/melee/gm/gm_17EB.c
@@ -162,7 +162,7 @@ bool fn_8017EDDC(void)
 #pragma push
 #pragma dont_inline on
 #endif
-void fn_8017EE40(int arg0_int)
+void fn_8017EE40(intptr_t arg0_int)
 {
     MatchEnd* arg0 = (MatchEnd*) arg0_int;
     struct StartMeleeRules* rules;
