+++ src/melee/gm/gmvs.c
@@ -928,7 +928,7 @@ MatchOutcome gm_GetMatchOutcome(void)
     return OUTCOME_NONE;
 }
 
-void fn_8016C46C(int arg0)
+void fn_8016C46C(intptr_t arg0)
 {
     if (gmVs_GetSceneController()->state.unk_9 != 0) {
         switch (gm_GetCurrentGameMode()) {
@@ -945,7 +945,7 @@ void fn_8016C46C(int arg0)
     }
 }
 
-static inline void fn_8016C46C_dontinline(int arg0)
+static inline void fn_8016C46C_dontinline(intptr_t arg0)
 {
     fn_8016C46C(arg0);
 }
@@ -1532,7 +1532,7 @@ void fn_8016D634(void)
             copied_dst->is_teams = controller.start.is_teams;
             copied_dst->outcome = tmp->state.match_result;
             gm_80166378(copied_dst);
-            fn_8016C46C_dontinline((int) copied_dst);
+            fn_8016C46C_dontinline((intptr_t) copied_dst);
             if (tmp->state.match_result != OUTCOME_NO_CONTEST &&
                 tmp->state.match_result != OUTCOME_RETRY)
             {
@@ -2107,7 +2107,7 @@ void gm_Scene_Vs_OnExit(void* user_data)
         data->xC.is_teams = controller.start.is_teams;
         data->xC.outcome = controller.state.match_result;
         gm_80166378(&data->xC);
-        fn_8016C46C_dontinline((int) &data->xC);
+        fn_8016C46C_dontinline((intptr_t) &data->xC);
         if (tmp->state.match_result != OUTCOME_NO_CONTEST &&
             tmp->state.match_result != OUTCOME_RETRY)
         {
